requirements.txt
streamlit
scikit-image
opencv-python-headless
Pillow
imutils
numpy
requests
